% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_modeling.R
\name{add_lag_column}
\alias{add_lag_column}
\title{add_lag_column}
\usage{
add_lag_column(dataset, lag_time = 30)
}
\arguments{
\item{dataset}{dataset}

\item{lag_time}{lag_time}
}
\value{
return
}
\description{
add_lag_column
}
