% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_modeling.R
\name{add_present_infection}
\alias{add_present_infection}
\title{add_present_infection}
\usage{
add_present_infection(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{
return
}
\description{
add_present_infection
}
