% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_modeling.R
\name{add_qpcr_times}
\alias{add_qpcr_times}
\title{add_qpcr_times}
\usage{
add_qpcr_times(dataset, qpcr_times)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{subject},
\code{time}, \code{allele}, and \code{present}.}

\item{qpcr_times}{A data.frame of qPCR positive only samples with columns
\code{subject} and \code{time}}
}
\value{
The \code{dataset} augmented with \code{subject} and \code{time} values for
qPCR positive only samples. In these rows, \code{present} will be set to 2 if
there was not already an \code{allele} with \code{present == 1} in the input
\code{dataset}.
}
\description{
Add times that were only qPCR positive (without sequencing information) as
2s in the \code{present} column. Times marked as qPCR positive that already
have sequencing information will be disregarded and left with their
sequencing values.
}
\examples{

dataset_in <- data.frame(allele = c('A', 'B', NA, NA, NA),
    subject = rep('A', 5),
    time = c(1, 1, 2, 3, 4))

qpcr_times <- data.frame(subject = rep('A', 2),
    time = c(2, 5))

dataset <- fill_in_dataset(dataset_in)
dataset <- add_qpcr_times(dataset, qpcr_times)

}
