% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_modeling.R
\name{add_time_gap}
\alias{add_time_gap}
\title{add_time_gap}
\usage{
add_time_gap(dataset, default = NA)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, and \code{present}.}

\item{default}{The reference time against which the first time point
is compared}
}
\value{
The dataset with a new column \code{time_gap} indicating
the time since the last observation.
}
\description{
Add a column with the time since the last observation. The \code{default} will
be used to define the \code{time_gap} for first visits.
}
\examples{

dataset_in <- data.frame(allele = c('A', 'B', NA, NA, NA),
    subject = rep('A', 5),
    time = c(1, 1, 2, 3, 4))

dataset <- fill_in_dataset(dataset_in)
dataset <- add_time_gap(dataset)

}
