% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_modeling.R
\name{add_treatment_column}
\alias{add_treatment_column}
\title{add_treatment_column}
\usage{
add_treatment_column(
  dataset,
  treatments,
  treatment_acute_start = 1,
  treatment_longitudinal_start = 10,
  verbose = TRUE
)
}
\arguments{
\item{dataset}{dataset}

\item{treatments}{treatments}

\item{treatment_acute_start}{treatment_acute_start}

\item{treatment_longitudinal_start}{treatment_longitudinal_start}

\item{verbose}{verbose}
}
\value{
return
}
\description{
add_treatment_column
}
