% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_modeling.R
\name{add_treatment_infection}
\alias{add_treatment_infection}
\title{add_treatment_infection}
\usage{
add_treatment_infection(
  dataset,
  treatments,
  treatment_acute_start = 1,
  treatment_longitudinal_start = 10,
  verbose = TRUE
)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, and \code{present_infection}}

\item{treatments}{A data.frame of treatments with columns \code{subject} and
\code{time}}

\item{treatment_acute_start}{Number of days after the treatment at which
to mark the start of the acute treatment phase, corresponding to when
variants are being cleared but might still be detected.}

\item{treatment_longitudinal_start}{Number of days after the treatment
at which to mark the start of the longitudinal treatment phase, corresponding
to when cleared variants will no longer be detected.}

\item{verbose}{Print an update ever 10 subjects since this function can
take some time.}
}
\value{
The dataset with new columns \code{treatment_acute_infection} and
\code{treatment_longitudinal_infection} (0/1) indicating that the infection was
treated in
the last \code{treatment_longitudinal_start} days (acute) or ever (longitudinal).
Columns for \code{persistent_infection} and \verb{lag_infection_[#]} will be set to 0
after treatment until the next infection.
}
\description{
Add columns with indicators of whether there is acute or longitudinal
(any previous) treatment for any allele
}
\examples{

dataset_in <- data.frame(allele = c('A', NA, NA, NA, 'B'),
                         subject = rep('A', 5),
                         time = c(1, 29, 30, 39, 50))
treatments <- data.frame(subject = c('A'),
                         time = c(29))

dataset <- fill_in_dataset(dataset_in)
dataset <- add_present_infection(dataset)
dataset <- add_persistent_column(dataset)
dataset <- add_lag_column(dataset)
dataset <- add_treatment_infection(dataset, treatments)

}
