% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_infections.R
\name{compute_total_new_COI}
\alias{compute_total_new_COI}
\title{compute_total_new_COI}
\usage{
compute_total_new_COI(dataset, method = "sum_then_max")
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, \code{present}, \code{locus}, \code{probability_new}, and
\code{probability_present} if using imputed data.}

\item{method}{Whether to sum the per-locus COI across all time points for
each locus and then take the maximum (\code{sum_then_max}) or take the maximum
per-locus COI at each time point across the loci and then sum the maxima
(\code{max_then_sum}).}
}
\value{
A dataframe with a \code{subject} column and a \code{new_COI} column
}
\description{
Compute the total new COI deduplicated across loci for each subject by
either summing across the loci and taking the max or maximizing the
diversity on each day and then summing the maxima.
}
\examples{

library(dplyr)
dataset_in <- data.frame(
    locus = rep(c(1, 2), each = 5),
    allele = c('A', 'B', 'B', NA, NA, 'A', 'A', 'B', NA, NA),
    subject = rep('A', 10),
    time = rep(c(1, 1, 5, 15, 44), 2)) \%>\%
    mutate(allele = interaction(allele, locus))

dataset <- fill_in_dataset(dataset_in)

dataset$probability_new <-
    determine_probabilities_simple(dataset)$probability_new

compute_total_new_COI(dataset, method = 'sum_then_max')
compute_total_new_COI(dataset, method = 'max_then_sum')

}
