% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_probabilities.R
\name{compute_total_new_molFOI}
\alias{compute_total_new_molFOI}
\title{compute_total_new_molFOI}
\usage{
compute_total_new_molFOI(dataset, method = "sum_then_max")
}
\arguments{
\item{dataset}{dataset}

\item{method}{method}
}
\value{
return
}
\description{
compute_total_new_molFOI
}
