% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_infections.R
\name{estimate_new_infections}
\alias{estimate_new_infections}
\title{estimate_new_infections}
\usage{
estimate_new_infections(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{
return
}
\description{
estimate_new_infections
}
