% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_probabilities.R
\name{impute_dataset}
\alias{impute_dataset}
\title{compute_total_new_molFOI}
\usage{
impute_dataset(dataset, n_imputations = 10, k = NULL, n_cores = 1, seed = 1)
}
\arguments{
\item{dataset}{dataset}

\item{n_imputations}{n_imputations}

\item{k}{k}

\item{n_cores}{n_cores}

\item{seed}{seed}
}
\value{
return
}
\description{
compute_total_new_molFOI
}
