% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_infections.R
\name{merge_probability_columns}
\alias{merge_probability_columns}
\title{merge_probability_columns}
\usage{
merge_probability_columns(dataset, cols_to_merge, threshold = 0.3)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, \code{present}, and \code{probability_present} if using imputed
data. This dataset should also contain columns with names given by
\code{cols_to_merge}.}

\item{cols_to_merge}{Which columns of probabilities in \code{dataset} to merge.}

\item{threshold}{Probabilities within \code{threshold} will be averaged
rather than manually selected.}
}
\value{
The \code{dataset} with the column \code{probability_new} given by
the average of the two columns when the two
columns are within \code{threshold} of each other and the manually chosen value
otherwise.
}
\description{
Take two columns of probabilities in the dataset and merge them by taking
the average if they are close or manually selecting when the probabilities
are discordant.
}
