% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dataset}
\alias{plot_dataset}
\title{plot_dataset}
\usage{
plot_dataset(dataset, output, treatments = NULL, height = 6, width = 8)
}
\arguments{
\item{output}{output}

\item{treatments}{treatments}

\item{height}{height}

\item{width}{width}

\item{subject}{subject}
}
\value{
return
}
\description{
plot_dataset
}
