% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_single_subject}
\alias{plot_single_subject}
\title{plot_single_subject}
\usage{
plot_single_subject(
  subject,
  dataset,
  treatments = NULL,
  estimated_new_infections = NULL,
  no_imputation = FALSE,
  output = NULL,
  height = 6,
  width = 8,
  highlight_index = NULL
)
}
\arguments{
\item{subject}{Subject to plot}

\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, and \code{present}. To plot imputations,
\code{probability_present} should be a column. To plot probabilities alleles
are new, \code{probability_new} should be a column. To display allele
prevalence, \code{prevalence} should be a column.}

\item{treatments}{A data.frame of treatments with columns \code{subject} and
\code{time}.}

\item{estimated_new_infections}{The result of \code{estimate_new_infections}
run on the dataset.}

\item{no_imputation}{Disregard qPCR-only times by setting times with
\code{present == 2} to \code{present = 0}.}

\item{output}{File to which the plot should be saved}

\item{height}{Height of the output plot}

\item{width}{Width of the output plot}

\item{highlight_index}{Index to highlight}
}
\value{
A plot object from ggplot2 showing the infection course for the
subject
}
\description{
Plot an infection time course for a single subject. The plot
shows when each allele was present (for each locus if applicable)
with annotations
for the probability the allele is new if \code{probability_new} is a column. Each
allele is its own row and color, and points are connected if the same allele
occurs repeatedly.
If \code{probability_present} is a column, imputed alleles are shown with an
opacity alpha equal to how often they were imputed.
If \code{probability_new} is a column, the pan-locus new
complexity of infection at each time is shown in a strip above the plot.
If an \code{estimated_new_infections} matrix or data frame is provided,
the plot title will include the estimated number of new infections.
Red dashes indicate a time point only known to be infection positive (i.e.,
no sequencing).
Blue dashes indicate treatment at that day if the \code{treatment} data frame
is provided.
If \code{probability_new} is a column, black dots indicate infections that are
new with >50\% probability, open dots indicate infections that are new with
<50\% probability.
If the dots are not annotated, the probabilities are >80\% or
<20\% respectively, and all other dots are annotated with their probabilities
of being new.
}
\examples{

dataset_in <- data.frame(allele = c('A', 'B', NA, NA, NA),
     subject = rep('A', 5),
     time = c(1, 1, 5, 15, 44))
treatments <- data.frame(subject = c('A', 'B'),
                         time = c(1, 29))

dataset <- fill_in_dataset(dataset_in)

dataset$probability_new <-
    determine_probabilities_simple(dataset)$probability_new

estimated_new_infections <- estimate_new_infections(dataset)

plot_single_subject('A', dataset, treatments, estimated_new_infections)

}
