% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_modeling.R
\name{add_present_infection}
\alias{add_present_infection}
\title{add_present_infection}
\usage{
add_present_infection(dataset)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, and \code{present.}}
}
\value{
The dataset with a new column \code{present_infection} (0/1) indicating
whether an infection is present at the time.
}
\description{
Add a column with an indicator of whether an infection is present at the
time.
}
\examples{

dataset_in <- data.frame(allele = c('A', 'B', NA, NA, NA),
    subject = rep('A', 5),
    time = c(1, 1, 2, 3, 4))

dataset <- fill_in_dataset(dataset_in)
dataset <- add_present_infection(dataset)

}
