% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_probabilities.R
\name{determine_probabilities_bayesian}
\alias{determine_probabilities_bayesian}
\title{determine_probabilities_bayesian}
\usage{
determine_probabilities_bayesian(
  dataset,
  infection_persistence_covariates = NULL,
  infection_general_covariates = NULL,
  alleles_persistence_covariates = NULL,
  chains = 1,
  parallel_chains = 1,
  iter_warmup = 500,
  iter_sampling = 500,
  refresh = 100,
  adapt_delta = 0.99,
  seed = 1,
  drop_out = FALSE
)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, and \code{present.}}

\item{infection_persistence_covariates}{A character vector of column names
in the dataset corresponding to covariates that increase only the
probability of an infection being present due to a persistent infection.}

\item{infection_general_covariates}{A character vector of column names
in the dataset corresponding to covariates that increase the
probability of a new infection being present}

\item{alleles_persistence_covariates}{A character vector of column names
in the dataset corresponding to covariates that increase only the
probability of an allele being present due to a persistent infection.}

\item{chains}{Number of chains for the Stan model}

\item{parallel_chains}{Number of parallel chains for the Stan model}

\item{iter_warmup}{Number of warm-up steps for the Stan model}

\item{iter_sampling}{Number of sampling steps for the Stan model}

\item{refresh}{Stan updates will be printed every \code{refresh} steps}

\item{adapt_delta}{Stan's \code{adapt_delta} parameter}

\item{seed}{Random seed for Stan}

\item{drop_out}{Whether to use a model that allows alleles to be called
persistent despite having never been previously observed due to sequencing
drop-out. This is recommended if more than 50\% of the alleles have dropped
out.}
}
\value{
A named list with (1) \code{probability_new} as a vector of the
probabilities the corresponding alleles in the dataset were from new
infections and (2) \code{fit} as the Stan model output.
}
\description{
Determine whether an allele is new or persistent using a Bayesian model.
A Bayesian model is parameterized with general and persistence covariates
to model the probability an infection (overall and allele-specific) is
present. The model is fit in Stan, and the parameters are used to estimate
the probability each allele is new.
}
