% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_probabilities.R
\name{determine_probabilities_clustering}
\alias{determine_probabilities_clustering}
\title{determine_probabilities_clustering}
\usage{
determine_probabilities_clustering(dataset, refresh = 100, seed = 1)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, and \code{present.}}

\item{refresh}{Stan updates will be printed every \code{refresh} steps}

\item{seed}{Random seed for Stan}
}
\value{
A named list with (1) \code{probability_new} as a vector of the
probabilities the corresponding alleles in the dataset were from new
infections and (2)
\code{fit} as the resulting parameters from the likelihood maximization
procedure.
}
\description{
Determine whether an allele is new or persistent using a clustering model.
Alleles are clustered according to co-occurrence, and the probability of
each allele occurring in each cluster and each cluster occurring at each time
is found through a likelihood maximization procedure. These probabilities
are used to estimate the probability an infection is new.
}
