% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_probabilities.R
\name{determine_probabilities_simple}
\alias{determine_probabilities_simple}
\title{determine_probabilities_simple}
\usage{
determine_probabilities_simple(dataset, n_lags = 3)
}
\arguments{
\item{dataset}{dataset}

\item{n_lags}{n_lags}
}
\value{
return
}
\description{
determine_probabilities_simple
}
