% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_covariates.R
\name{estimate_drop_out}
\alias{estimate_drop_out}
\title{estimate_drop_out}
\usage{
estimate_drop_out(dataset, allowed_ordered_time_difference = 6)
}
\arguments{
\item{dataset}{dataset}

\item{allowed_ordered_time_difference}{allowed_ordered_time_difference}
}
\value{
return
}
\description{
estimate_drop_out
}
