% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_infections.R
\name{estimate_new_infections}
\alias{estimate_new_infections}
\title{estimate_new_infections}
\usage{
estimate_new_infections(dataset, imputation_mat = NULL, probability_mat = NULL)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, \code{present} (0/1), and \code{probability_new}.}

\item{imputation_mat}{If using imputations, the output of the function
\code{impute_dataset}: a
matrix with an equal number of rows to \code{dataset} and one column per
imputed dataset.}

\item{probability_mat}{If using imputations, a matrix with the same
dimensions as \code{imputation_mat}
with one column per imputed dataset containing the probability the allele
was new if present.}
}
\value{
A data frame with row names corresponding to the subjects and
one column per imputed dataset (or a single column if not using
imputations).
}
\description{
Compute the total new infection events for each subject by identifying
peaks in the pan-locus new alleles.
}
\examples{

library(dplyr)
dataset_in <- data.frame(
    locus = rep(c(1, 2), each = 5),
    allele = c('A', 'B', 'B', NA, NA, 'A', 'A', 'B', NA, NA),
    subject = rep('A', 10),
    time = rep(c(1, 1, 5, 15, 44), 2)) \%>\%
    mutate(allele = interaction(allele, locus))

dataset <- fill_in_dataset(dataset_in)

dataset$probability_new <-
    determine_probabilities_simple(dataset)$probability_new

estimate_new_infections(dataset)

}
