% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dataset}
\alias{plot_dataset}
\title{plot_dataset}
\usage{
plot_dataset(
  dataset,
  treatments = NULL,
  estimated_new_infections = NULL,
  no_imputation = FALSE,
  output = NULL,
  height = 6,
  width = 8
)
}
\arguments{
\item{dataset}{A complete longitudinal dataset with columns \code{allele},
\code{subject}, \code{time}, and \code{present}. To plot imputations,
\code{probability_present} should be a column. To plot probabilities alleles
are new, \code{probability_new} should be a column.}

\item{treatments}{A data.frame of treatments with columns \code{subject} and
\code{time}.}

\item{estimated_new_infections}{The result of \code{estimate_new_infections}
run on the dataset.}

\item{no_imputation}{Disregard qPCR-only times by setting times with
\code{present == 2} to \code{present = 0}.}

\item{output}{Folder to which the plots should be saved}

\item{height}{Height of the output plot}

\item{width}{Width of the output plot}
}
\value{
A list of plot object from ggplot2 showing each per-subject
infection course. These are also saved by subject name in the \code{output}
folder.
}
\description{
Compute allele prevalences and then call \code{plot_single_subject} for every
subject in a dataset and save the results.
}
\examples{

library(dplyr)
library(dinemites)
dataset_in <- data.frame(allele = c('A', 'B', NA, NA, NA),
     subject = rep('A', 5),
     time = c(1, 1, 5, 15, 44))
treatments <- data.frame(subject = c('A', 'B'),
                         time = c(1, 29))

dataset <- fill_in_dataset(dataset_in)

dataset$probability_new <-
    determine_probabilities_simple(dataset)$probability_new

plot_dataset(dataset, treatments)

}
