% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_single_subject}
\alias{plot_single_subject}
\title{plot_single_subject}
\usage{
plot_single_subject(
  subject,
  dataset,
  output = NULL,
  treatments = NULL,
  height = 6,
  width = 8
)
}
\arguments{
\item{subject}{subject}

\item{dataset}{dataset}

\item{output}{output}

\item{treatments}{treatments}

\item{height}{height}

\item{width}{width}
}
\value{
return
}
\description{
plot_single_subject
}
